# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API Sistem Informasi Akademik Tanya UNPRA
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List pengumuman (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: kategori
          in: query
          schema:
            type: string
            enum: [Akademik, Kemahasiswaan, Keuangan, Umum, Beasiswa]
        - name: isActive
          in: query
          schema:
            type: string
            enum: ["true", "false"]
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Daftar pengumuman
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementsResponse"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Buat pengumuman baru (admin & dosen)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnnouncementRequest"
      responses:
        "201":
          description: Pengumuman berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementResponse"
        "400":
          description: Data tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /announcements/{id}:
    get:
      operationId: getAnnouncement
      tags: [announcements]
      summary: Detail pengumuman (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Detail pengumuman
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementResponse"
        "404":
          description: Pengumuman tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update pengumuman (admin edit semua, dosen edit milik sendiri)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnnouncementRequest"
      responses:
        "200":
          description: Pengumuman berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementResponse"
        "403":
          description: Dosen hanya bisa edit pengumuman sendiri
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Pengumuman tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Hapus pengumuman (admin semua, dosen milik sendiri)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Pengumuman berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "403":
          description: Akses ditolak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Pengumuman tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /academic-calendar:
    get:
      operationId: listAcademicCalendar
      tags: [academic-calendar]
      summary: List event kalender akademik (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: tahunAjaran
          in: query
          schema: { type: string }
        - name: tipe
          in: query
          schema:
            type: string
            enum: [UTS, UAS, Libur, Registrasi, KRS, Wisuda, Lainnya]
        - name: from
          in: query
          schema:
            type: string
            format: date
          description: Filter event mulai dari tanggal ini (YYYY-MM-DD)
        - name: to
          in: query
          schema:
            type: string
            format: date
          description: Filter event sampai tanggal ini (YYYY-MM-DD)
      responses:
        "200":
          description: Daftar event kalender
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcademicCalendarListResponse"
    post:
      operationId: createCalendarEvent
      tags: [academic-calendar]
      summary: Tambah event kalender akademik (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarEventRequest"
      responses:
        "201":
          description: Event berhasil dibuat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEventResponse"
        "400":
          description: Data tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Akses ditolak
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /academic-calendar/{id}:
    get:
      operationId: getCalendarEvent
      tags: [academic-calendar]
      summary: Detail event kalender (semua role)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Detail event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEventResponse"
        "404":
          description: Event tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateCalendarEvent
      tags: [academic-calendar]
      summary: Update event kalender (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCalendarEventRequest"
      responses:
        "200":
          description: Event berhasil diperbarui
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEventResponse"
        "404":
          description: Event tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCalendarEvent
      tags: [academic-calendar]
      summary: Hapus event kalender (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Event berhasil dihapus
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Event tidak ditemukan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [page, limit, total, totalPages]
...
    Announcement:
      type: object
      properties:
        id:
          type: string
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: string
          enum: [Akademik, Kemahasiswaan, Keuangan, Umum, Beasiswa]
        authorId:
          type: string
          nullable: true
        authorName:
          type: string
          nullable: true
        isActive:
          type: boolean
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, judul, konten, kategori, isActive, publishedAt, createdAt, updatedAt]
    AnnouncementResponse:
      type: object
      properties:
        announcement:
          $ref: "#/components/schemas/Announcement"
      required: [announcement]
    AnnouncementsResponse:
      type: object
      properties:
        announcements:
          type: array
          items:
            $ref: "#/components/schemas/Announcement"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [announcements, pagination]
    CreateAnnouncementRequest:
      type: object
      properties:
        judul:
          type: string
          minLength: 5
          maxLength: 300
        konten:
          type: string
          minLength: 10
          maxLength: 10000
        kategori:
          type: string
          enum: [Akademik, Kemahasiswaan, Keuangan, Umum, Beasiswa]
          default: Umum
        isActive:
          type: boolean
          default: true
        publishedAt:
          type: string
          format: date-time
      required: [judul, konten]
    UpdateAnnouncementRequest:
      type: object
      properties:
        judul:
          type: string
        konten:
          type: string
        kategori:
          type: string
          enum: [Akademik, Kemahasiswaan, Keuangan, Umum, Beasiswa]
        isActive:
          type: boolean
        publishedAt:
          type: string
          format: date-time
    CalendarEvent:
      type: object
      properties:
        id:
          type: string
        namaEvent:
          type: string
        tanggalMulai:
          type: string
          format: date
        tanggalSelesai:
          type: string
          format: date
        tipe:
          type: string
          enum: [UTS, UAS, Libur, Registrasi, KRS, Wisuda, Lainnya]
        deskripsi:
          type: string
          nullable: true
        tahunAjaran:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, namaEvent, tanggalMulai, tanggalSelesai, tipe, tahunAjaran, createdAt, updatedAt]
    CalendarEventResponse:
      type: object
      properties:
        event:
          $ref: "#/components/schemas/CalendarEvent"
      required: [event]
    AcademicCalendarListResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/CalendarEvent"
        pagination:
          $ref: "#/components/schemas/Pagination"
      required: [events, pagination]
    CreateCalendarEventRequest:
      type: object
      properties:
        namaEvent:
          type: string
          minLength: 3
          maxLength: 200
        tanggalMulai:
          type: string
          format: date
          description: Format YYYY-MM-DD
        tanggalSelesai:
          type: string
          format: date
          description: Format YYYY-MM-DD
        tipe:
          type: string
          enum: [UTS, UAS, Libur, Registrasi, KRS, Wisuda, Lainnya]
        deskripsi:
          type: string
          maxLength: 1000
        tahunAjaran:
          type: string
      required: [namaEvent, tanggalMulai, tanggalSelesai, tipe, tahunAjaran]
    UpdateCalendarEventRequest:
      type: object
      properties:
        namaEvent:
          type: string
        tanggalMulai:
          type: string
          format: date
        tanggalSelesai:
          type: string
          format: date
        tipe:
          type: string
          enum: [UTS, UAS, Libur, Registrasi, KRS, Wisuda, Lainnya]
        deskripsi:
          type: string
        tahunAjaran:
          type: string
...